Move emit output collector to it's own file
